llm_deployment_project/requirements.txt
transformers
torch
accelerate
bitsandbytes
sentencepiece
fastapi
uvicorn[standard]
pydantic
python-dotenv
locust